\name{PCA}
\alias{PCA}

%- Also NEED an '\alias' for EACH other topic documented here.
\title{ksIRT - kernel smoothing in Item Response Theory
}
\description{
Returns a vector with each subjects expected test score.
}
\usage{
PCA(x,...)

}
%- maybe also 'usage' for other objects documented here.
\arguments{

\item{x}{
a \code{ksIRT} object to be analyzed.
}
\item{...}{
other arguments to pass to the \code{prcomp} function from the \code{stats} package.
}
}


\references{
Ramsay, J.O. (2000). 
\pkg{TestGraf}: A program for the graphical analysis of multiple choice test and questionnaire data. 
   
 Silverman, B.W. (1986). 
 \emph{Density Estimation for Statistics and Data Analysis}. 
 Chapman & Hall, London.
}




