\name{subjscore}
\alias{subjscore}

%- Also NEED an '\alias' for EACH other topic documented here.
\title{ksIRT - kernel smoothing in Item Response Theory
}
\description{
Returns a vector of scores for each subject. 
}
\usage{
subjscore(x)

}
%- maybe also 'usage' for other objects documented here.
\arguments{

\item{x}{
a \code{ksIRT} object to be analyzed.
}


}


\references{
Ramsay, J.O. (2000). 
\pkg{TestGraf}: A program for the graphical analysis of multiple choice test and questionnaire data. 
   
 Silverman, B.W. (1986). 
 \emph{Density Estimation for Statistics and Data Analysis}. 
 Chapman & Hall, London.
}




